+++ assets/javascripts/kitten/index.js
@@ -243,6 +243,7 @@ export { KkbbFlashIllustration } from './components/graphics/illustrations/kkbb-
 export { KissKissBankBankLogo } from './components/graphics/logos/kisskissbankbanklogo'
 export { LendopolisLogo } from './components/graphics/logos/lendopolis-logo'
 export { LendopolisLogoOnDark } from './components/graphics/logos/lendopolis-logo/on-dark'
+export { LendopolisLogoVerticalOnDark } from './components/graphics/logos/lendopolis-logo/vertical-on-dark'
 export { GoodeedLogo } from './components/graphics/logos/goodeedlogo'
 export { MaisonDeCrowdfundingLogo } from './components/graphics/logos/mdc'
 export { MaisonDeCrowdfundingLogoOnDark } from './components/graphics/logos/mdc/on-dark'