+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: `DeskLayout`: Fix menu column width.
+- Fix: `ImageDropUploader`: Add background color.
 
 ## [6.16.0] - 2022-01-05