+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `ContributionCard`: Fix image display.
 - Feature: `Tag`: Add `size` prop defaults to `regular`.
 - Fix: `Tag`:  New UI - colors style.
 - Feature: Logos: Add `LendopolisLogoVerticalOnDark` logo.
@@ -12,7 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [6.16.0] - 2022-01-05
 
-Features: 
+Features:
+
 - `ProjectCard`: Add `ProjectCard.Sticker` subcomponent.
 - `ProjectCard`: Deprecate `sticker` prop.
 - `Utilities`: Add `k-u-transform-capitalize` utility class.